requirements.txt
music_tag==0.4.3
python-dotenv==1.0.0
Requests==2.31.0
yt_dlp==2023.7.6
